ユーザー関連と問い合わせ関連のVueページをそれぞれuserとcontactというフォルダにまとめる
