+++ frontend/src/components/page/contact/ContactDetail.vue
@@ -1,15 +1,15 @@
 <script setup>
 import { ref, onMounted } from "vue";
 import { useRoute, useRouter } from "vue-router";
-import { axiosInstance, setupInterceptors } from "../../const/axios.js";
+import { axiosInstance, setupInterceptors } from "../../../const/axios.js";
 import { useToast } from "primevue/usetoast";
-import { toastService } from "../../const/toast.js";
+import { toastService } from "../../../const/toast.js";
 
 const toast = useToast();
 const toastNotifications = new toastService(toast);
 
-import Header from "../layout/Header.vue";
-import TabMenu from "../layout/TabMenu.vue";
+import Header from "../../layout/Header.vue";
+import TabMenu from "../../layout/TabMenu.vue";
 
 const route = useRoute();
 const router = useRouter();
