+++ frontend/src/components/page/user/UserList.vue
@@ -1,14 +1,14 @@
 <script setup>
 import { ref, onMounted } from "vue";
 import { useRouter, useRoute, onBeforeRouteUpdate } from "vue-router";
-import { axiosInstance, setupInterceptors } from "../../const/axios.js";
+import { axiosInstance, setupInterceptors } from "../../../const/axios.js";
 
-import TabMenu from "../layout/TabMenu.vue";
-import Header from "../layout/Header.vue";
-import UserCard from "../layout/UserCard.vue";
-import ListPage from "../layout/ListPage.vue";
-import SearchButton from "../atom/SearchButton.vue";
-import ResultEmpty from "../atom/ResultEmpty.vue";
+import TabMenu from "../../layout/TabMenu.vue";
+import Header from "../../layout/Header.vue";
+import UserCard from "../../layout/UserCard.vue";
+import ListPage from "../../layout/ListPage.vue";
+import SearchButton from "../../atom/SearchButton.vue";
+import ResultEmpty from "../../atom/ResultEmpty.vue";
 
 const route = useRoute();
 const router = useRouter();
@@ -19,7 +19,7 @@ const isDisplayOnlySupport = ref(false);
 const isLogin = ref(false);
 const searchResult = ref([]);
 const pageCount = ref(1);
-const pageNum = ref(1);
+const page = ref(1);
 
 onMounted(() => {
   setQuery(route.query.keyword, route.query.onlySupport, route.query.page);
@@ -135,7 +135,7 @@ const userSelect = (item) => {
       <div class="mt-12">
         <ListPage
           :pageCount="pageCount"
-          v-model="pageNum"
+          v-model="page"
           @changePage="updatePaginateItems"
         />
       </div>
