+++ frontend/src/router/index.js
@@ -1,8 +1,8 @@
 import { createRouter, createWebHistory } from 'vue-router';
 import Home from '/src/components/page/Home.vue';
-import AccountInteroduction from '/src/components/page/AccountIntroduction.vue';
-import Signup from '/src/components/page/Signup.vue';
-import Login from '/src/components/page/Login.vue';
+import Contact from '/src/components/page/contact/Contact.vue';
+import ContactList from '/src/components/page/contact/ContactList.vue';
+import ContactDetail from '/src/components/page/contact/ContactDetail.vue';
 import AddRecord from '/src/components/page/record/AddRecord.vue';
 import EditRecord from '/src/components/page/record/EditRecord.vue';
 import RecordDetail from '/src/components/page/record/RecordDetail.vue';
@@ -11,21 +11,21 @@ import KnowledgeList from '/src/components/page/knowledge/KnowledgeList.vue';
 import AddKnowledge from '/src/components/page/knowledge/AddKnowledge.vue';
 import EditKnowledge from '/src/components/page/knowledge/EditKnowledge.vue';
 import KnowledgeDetail from '/src/components/page/knowledge/KnowledgeDetail.vue';
+import UserProfile from '/src/components/page/user/UserProfile.vue';
+import EditProfile from '/src/components/page/user/EditProfile.vue';
+import UserList from '/src/components/page/user/UserList.vue';
+import MyPage from '/src/components/page/user/MyPage.vue';
 
-
-import UserProfile from '/src/components/page/UserProfile.vue';
-import EditProfile from '/src/components/page/EditProfile.vue';
+import AccountInteroduction from '/src/components/page/AccountIntroduction.vue';
+import Signup from '/src/components/page/Signup.vue';
+import Login from '/src/components/page/Login.vue';
 import PasswordEdit from '/src/components/page/PasswordEdit.vue';
 import MailAddressEdit from '/src/components/page/MailAddressEdit.vue';
-import Contact from '/src/components/page/Contact.vue';
-import ContactList from '/src/components/page/ContactList.vue';
-import ContactDetail from '/src/components/page/ContactDetail.vue';
 import NotFound from '/src/components/page/PageNotFoundError.vue';
-import UserList from '/src/components/page/UserList.vue';
 import ConfirmAccountDelete from '/src/components/page/ConfirmAccountDelete.vue';
 import SupportList from '/src/components/page/SupportList.vue';
 import SupporterList from '/src/components/page/SupporterList.vue';
-import MyPage from '/src/components/page/MyPage.vue';
+
 
 const routes = [
   {
